requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
faiss-cpu>=1.9.0
numpy>=1.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0

